Cleanup Update
Utils:
-removed unused variables from several functions
-alphabetized the functions in the cpp file
Simulation class:
-changed N_events event counter from int to long int data type to handle
larger numbers of events that can occur when running long simulations
-updated getN_events_executed function to return a long int

+++ Simulation.cpp
@@ -210,7 +210,7 @@ int Simulation::getNumSites(){
     return Length*Width*Height;
 }
 
-int Simulation::getN_events_executed(){
+long int Simulation::getN_events_executed(){
     return N_events_executed;
 }
